Add a variable for force linking the oc tool (#4)
Currently the link task will fail if the oc binary already exists
before attempting to create the link. If the 'force' option is set
to 'yes' then this error will not happen and the original file will
be overwritten.

To keep the original behaviour this option defaults to 'no'. Meaning
there is no issue with backwards compatibility.