Allow for oc unarchive to be forced
Currently when the oc binary already exists the unarchiving of the
downloaded version will be skipped. This is due to the 'creates'
field in the unarchive module being set.

This change adds a variable named force_oc_unarchive that will
allow to force the unarchiving of the downloaded oc tool
regardless of whether it currently exists or not. The new
variable is set to 'no' by default to keep backwards
compatibility.

The ternary filter used in this change was introduced in Ansible
1.9.